Refactor HistoricalStorage into smaller submodules
